Update. Перенос Popup с информацией инициативы с возможностью заказа в отдельный компонент
